seleniumMASS/requirements.txt
streamlit>=1.24.0
selenium>=4.11.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
webdriver_manager>=4.0.0
python-multipart>=0.0.6
pandas>=2.0.0
langchain>=0.0.300
langchain-community>=0.0.10
langchain-core>=0.0.10
ollama>=0.1.0
openpyxl>=3.1.0
